+++ hutool/cron/__init__.py
@@ -1,4 +1,5 @@
 from .cron import CronUtil
 from .cron_pattern import CronPattern
+from .cron_validator import CronValidator
 
-__all__ = ["CronPattern", "CronUtil"]
+__all__ = ["CronPattern", "CronUtil", "CronValidator"]